+++ tasks_functions.ino
@@ -41,9 +41,9 @@ void warm_cool(float setPoint)
 void read_settings_from_pref()
 {
 	xSemaphoreTake(pref_mutex, portMAX_DELAY);
-	man_mode_set_p = pref.getInt("man_mode_set_p");
-	day_set_p = pref.getInt("day_set_p");    
-	night_set_p = pref.getInt("night_set_p");    
+	man_mode_set_p = pref.getFloat("man_mode_set_p");
+	day_set_p = pref.getFloat("day_set_p");
+	night_set_p = pref.getFloat("night_set_p");
 	max_water_temp = pref.getInt("max_water_temp");    
 	min_water_temp = pref.getInt("min_water_temp");    
 	guard_mode = pref.getInt("guard_mode");    
@@ -370,9 +370,9 @@ void write_setting_to_pref(void *pvParameters)
 	while (true)
 	{
 		xSemaphoreTake(pref_mutex, portMAX_DELAY);
-		pref.putInt("man_mode_set_p", man_mode_set_p);
-		pref.putInt("day_set_p", day_set_p);    
-		pref.putInt("night_set_p", night_set_p);    
+		pref.putFloat("man_mode_set_p", man_mode_set_p);
+		pref.putFloat("day_set_p", day_set_p);    
+		pref.putFloat("night_set_p", night_set_p);    
 		pref.putInt("max_water_temp", max_water_temp);    
 		pref.putInt("min_water_temp", min_water_temp);
 		pref.putInt("guard_mode", guard_mode);
